+++ react/components/ShippingOptions.tsx
@@ -1,14 +1,20 @@
 import React, { FunctionComponent, Fragment } from 'react'
 import { Radio } from 'vtex.styleguide'
 import ShippingInfo from './ShippingInfo'
+import DeliveryOptionsAvailability from './DeliveryOptionsAvailability'
 import { slugify } from '../utils/slugify'
+import { FormattedMessage } from 'react-intl'
 
 interface CustomProps {
+  numberOfItems: number
+  numberOfUnavailableItems: number
   options: DeliveryOption[]
   selectDeliveryOption: (option: string) => void
 }
 
 const ShippingOptions: FunctionComponent<CustomProps> = ({
+  numberOfItems,
+  numberOfUnavailableItems,
   options,
   selectDeliveryOption,
 }) => {
@@ -16,8 +22,26 @@ const ShippingOptions: FunctionComponent<CustomProps> = ({
     selectDeliveryOption(event.currentTarget.value)
   }
 
+  const allItemsUnavailable = numberOfUnavailableItems === numberOfItems
+
   return (
     <Fragment>
+      <DeliveryOptionsAvailability
+        numberOfItems={numberOfItems}
+        numberOfUnavailableItems={numberOfUnavailableItems}
+      />
+      {numberOfUnavailableItems > 0 && !allItemsUnavailable && (
+        <div className="fw5 mt5 mb5">
+          <FormattedMessage
+            id="store/shipping-calculator.optionsForRemainingProducts"
+            values={{
+              numberOfOptions: options.length,
+              numberOfItems: numberOfItems - numberOfUnavailableItems,
+            }}
+          />
+        </div>
+      )}
+
       {options &&
         options.length > 0 &&
         options.map((option, i) => {